+++ server/src/test/documentFormating.test.ts
@@ -1289,10 +1289,28 @@ describe('Document formating', () => {
 			);
 		});
 
-		test('hex with 0XD', async () => {
+		test('hex with 0XD > on same line', async () => {
 			const documentText = '/ {\n\tprop1 = <0XD>;';
 			const newText = await getNewText(documentText);
 			expect(newText).toEqual('/ {\n\tprop1 = <0XD>;');
+		});
+
+		test('hex with 0XD > on new line', async () => {
+			const documentText = '/ {\n\tprop1 = <0XD\n>;';
+			const newText = await getNewText(documentText);
+			expect(newText).toEqual('/ {\n\tprop1 = <0XD>;');
+		});
+
+		test('hex with 0XD with line comment > on new line', async () => {
+			const documentText = '/ {\n\tprop1 = <0XD // test\n>;';
+			const newText = await getNewText(documentText);
+			expect(newText).toEqual('/ {\n\tprop1 = <0XD // test\n\t>;');
+		});
+
+		test('hex with 0XD with block comment > on new line', async () => {
+			const documentText = '/ {\n\tprop1 = <0XD /* test */\n>;';
+			const newText = await getNewText(documentText);
+			expect(newText).toEqual('/ {\n\tprop1 = <0XD /* test */ >;');
 		});
 
 		test('empty array value with spaces', async () => {