+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.5.6] - 2025-10-09
+## [0.5.6] - 2025-10-08
 
 - Fixed an issue with parsing block comment after line comment when block comment has "'" or '"' inside of it.
 - Fixed an issue where logical not '!' in expression was not supported properly.